Seed the DB.  Seeded 3 Exercises, 3 Goals, and 14 Exercise Records

+++ WeWorkOut/Controllers/ExerciseRecordsController.cs
@@ -102,7 +102,7 @@ namespace WeWorkOut.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!ExerciseRecordExists(exerciseRecord.ExerciseRecordID))
+                    if (!ExerciseRecordExists((int)exerciseRecord.ExerciseRecordID))
                     {
                         return NotFound();
                     }
